NAS-106239 / 12.0 / Do not remove automatically configured IPv6 default gateway (by themylogin)

+++ src/middlewared/middlewared/plugins/network.py
@@ -13,8 +13,10 @@ from collections import defaultdict
 import contextlib
 import ipaddress
 import itertools
+import os
 import platform
 import re
+import signal
 import socket
 import subprocess
 
@@ -29,6 +31,9 @@ ANNOUNCE_SRV = {
     'netbios': 'nmbd',
     'wsd': 'wsdd'
 }
+
+RE_RTSOLD_INTERFACE = re.compile(r'Interface (.+)')
+RE_RTSOLD_NUMBER_OF_VALID_RAS = re.compile(r'number of valid RAs: ([0-9]+)')
 
 
 class NetworkConfigurationModel(sa.Model):
@@ -1050,7 +1055,7 @@ class InterfaceService(CRUDService):
                     if mtu and mtu > (parent_iface.get('mtu') or 1500):
                         verrors.add(
                             f'{schema_name}.mtu',
-                            f'VLAN MTU cannot be bigger than parent interface.',
+                            'VLAN MTU cannot be bigger than parent interface.',
                         )
 
         failover_licensed = False
@@ -1991,8 +1996,22 @@ class RouteService(Service):
         elif routing_table.default_route_ipv6:
             # If there is no gateway in database but one is configured
             # remove it
-            self.logger.info('Removing IPv6 default route')
-            routing_table.delete(routing_table.default_route_ipv6)
+            interface = routing_table.default_route_ipv6.interface
+            autoconfigured_interface = await self.middleware.call(
+                'datastore.query', 'network.interfaces', [
+                    ['int_interface', '=', interface],
+                    ['int_ipv6auto', '=', True],
+                ]
+            )
+            remove = False
+            if not autoconfigured_interface:
+                self.logger.info('Removing IPv6 default route as there is no IPv6 autoconfiguration')
+                remove = True
+            elif not await self.middleware.call('route.has_valid_router_announcements', interface):
+                self.logger.info('Removing IPv6 default route as IPv6 autoconfiguration has not succeeded')
+                remove = True
+            if remove:
+                routing_table.delete(routing_table.default_route_ipv6)
 
     @accepts(Str('ipv4_gateway'))
     def ipv4gw_reachable(self, ipv4_gateway):
@@ -2010,6 +2029,54 @@ class RouteService(Service):
                         ipv4_nic = ipaddress.IPv4Interface(nic_address)
                         if ipaddress.ip_address(ipv4_gateway) in ipv4_nic.network:
                             return True
+        return False
+
+    @private
+    async def has_valid_router_announcements(self, interface):
+        rtsold_dump_path = '/var/run/rtsold.dump'
+
+        try:
+            with open('/var/run/rtsold.pid') as f:
+                rtsold_pid = int(f.read().strip())
+        except (FileNotFoundError, ValueError):
+            self.logger.warning('rtsold pid file does not exist')
+            return False
+
+        with contextlib.suppress(FileNotFoundError):
+            os.unlink(rtsold_dump_path)
+
+        try:
+            os.kill(rtsold_pid, signal.SIGUSR1)
+        except ProcessLookupError:
+            self.logger.warning('rtsold is not running')
+            return False
+
+        for i in range(10):
+            await asyncio.sleep(0.2)
+            try:
+                with open(rtsold_dump_path) as f:
+                    dump = f.readlines()
+                    break
+            except FileNotFoundError:
+                continue
+        else:
+            self.logger.warning('rtsold has not dumped status')
+            return False
+
+        current_interface = None
+        for line in dump:
+            line = line.strip()
+
+            m = RE_RTSOLD_INTERFACE.match(line)
+            if m:
+                current_interface = m.group(1)
+
+            if current_interface == interface:
+                m = RE_RTSOLD_NUMBER_OF_VALID_RAS.match(line)
+                if m:
+                    return int(m.group(1)) > 0
+
+        self.logger.warning('Have not found %s status in rtsold dump', interface)
         return False
